% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_segment.R
\name{cell_segment}
\alias{cell_segment}
\title{Cell segmentation}
\usage{
cell_segment(img_num = y, img_w = 1280, img_h = 1024, lambda_adj = 10^1)
}
\description{
Cell segmentation function of uploaded and converted images. Users can adjust
the provided variables to improve cell segementation.
}
