% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_process.R
\name{img_process}
\alias{img_process}
\title{Image Process}
\usage{
img_process(
  img_num = 1,
  contrast_adj = 0.2,
  bright_adj = 0.1,
  increase = TRUE,
  thresh_adj = 0.28,
  w = 50,
  h = 50,
  offset = 0.001,
  area_thresh = 250,
  tolerance = 0.8,
  ext = 3
)
}
\description{
Initial conversion of uploaded images using tools from the *cellcount* package.
Users can adjust various parameters to meet the needs that the image requires for optimal
cell segmentation.
}
